Updated integration tests for Subtensor
